backend/pyproject.toml
[project]
name = "postmail-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.8.0",
    "sentence-transformers>=2.2.0",
    "newspaper4k>=0.9.0",
    "beautifulsoup4>=4.12.0",
    "youtube-transcript-api>=0.6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
